pyproject.toml
[project]
name = "lunch-menu-bot"
version = "0.1.0"
description = "A Discord bot that fetches the daily lunch menu"
readme = "README.md"
requires-python = "==3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "discord-py>=2.4.0",
    "openai>=1.52.2",
    "pre-commit>=4.0.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
